Prevents custom post types from being published to Discourse (if needed)
This forcibly prevents custom post types (for example, as used by some wordpress plugins) from being published to Discourse.

It's a little heavy-handed, because it essentially blocks _all_ custom posts from being published, which might actually be desired behaviour in some situations. 

It might also not be necessary, for example I just tried using the "Liveblogging" plugin without this fix, which didn't actually publish anything to Discourse even with the option checked, which may mean it's down to how the custom posts are implemented.

I leave it to someone with more wp plugin experience to chime in, and I'll do some additional testing.