Slight improvement to scheduling posts
I don't think the `$old_status` is needed. If we use the `transition_post_status` hook just to check for the `$new_status` to be `publish` that should be fine. That will then allow an unchecked "Publish to Discourse" box to be checked after the post is published and it will post to Discourse without needing to change the post status.

The reason the `$old_status` was being used is I thought that a post might get published to Discourse multiple times if you were to simply update the post, but that doesn't appear to be the case.