Allow showing existing WP comments under Discourse
This patch is my answer to what to do about enabling Discourse comments for existing posts/pages that have a lot of valuable pre-existing WP comments on them. The prior behavior of the plugin makes you choose to either show Discourse comments and hide the old WP comments OR show the WP comments and not show the Discourse comments. This patch lets you choose to show the existing WP comments as a "historical archive" beneath the new Discourse comments.