Allow redirect to Discourse without login
if using SSO, the link that wp-discourse generates that sends the user to the Discourse thread to comment on the WP post ("continue the discussion in our forum"), forces the user to login before they are redirected to the Discourse thread.

i decided that i wanted to have the link redirect straight to the thread without an intervening login page. if the user wants to post when they get there, they can click the "login to reply" button in Discourse.

this patch adds a setting in the admin that allows this behavior (but leaves the default as-is).